admin_portal/requirements.txt
# Django
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0

# Database
psycopg2-binary==2.9.10

# Redis & Celery
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Structured logging
structlog==24.4.0

# Utilities
python-dotenv==1.0.1

# Linting & type checking
ruff==0.8.6
mypy==1.14.1
django-stubs==5.1.1
